requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
crispy-bootstrap5==0.7
cryptography==45.0.7
Django==5.2.6
django-allauth==65.11.1
django-crispy-forms==2.1
idna==3.10
Pillow==10.0.1
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.5
sqlparse==0.5.3
stripe==7.8.0
typing_extensions==4.15.0
urllib3==2.5.0
